Moved Cpg class and the loading classes to codepropertygraph project.
With the exception of the top level loading class CpgLoader.
The reason for this is enable us to load a CPG represented by the Cpg
class without requiring the query-primitives or semantic cpg classes.

The implicit import extension classes also moved to codepropertygraph
because the are extended by the Cpg class. They are still in the
query-primitives name space in order to make the changes not all at
once.