Close resources after use - tests (#227)
* Close resources after use - tests

While running sbt/test continuously I saw a steady increase in the used
heap. This is because CPGs weren't being closed at all and with ondisk
overflow that is actually problematic.

While we might turn off ondisk overflow for tests and such it is still a
good idea to always close the resources in a clean way.

The change essentially cleans up the way we create/use test fixtures to
ensure that resources are always closed. With the change, running tests
continuously kept heap constant and low.

* Don't use `autoclose`

Instead of explicitly calling `autoclose`, let's use closures in
fixtures to ensure that they correctly close resources.
This should simplify tests.

* scalafmt

* nit