Simplify CpgNodeMethods.
In flow pretty printing, we were using `cpgNode.code`, which looked like
it was just accessing a `code` field but actually read a different field,
depending on the type of the CfgNode. This was misleading. In addition,
`code` was implemented by using the enrich-my-library pattern and the
visitor pattern, making it quite hard to understand what was going on.

What we are really doing here is just obtaining a string representation
for each CfgNode. I therefore replaced the old CfgNodeMethods with a
static method called `repr` that uses a `match`. We therefore no longer
give the impression that we are just reading the `code` field, avoiding
confusion. This is used in pretty printing of flows only and should not
be a performance issue.