RIP `Labels` type parameter (#330)
The `Labels` type parameter allowed us to refer back to previous steps
in a type-safe manner, using the `as/select` construct. This was neat,
but we never really used this feature. Since managing the type
parameter is quite a lot of code, which is especially hard to read for
newcomers to this codebase and/or scala, we're now dropping it altogether.