size/count and is steps
Triggered https://github.com/ShiftLeftSecurity/codescience/issues/2995
it became obvious that we're missing a few basic wrapper steps. These
are easy to add though.

`is(P)` is a bit of a leaky abstraction, but not too bad IMO.

n.b. independent of this we can still rename `filter` ->
`filterOnTrav` and `filterOnEnd` -> `filter`. Downside would be that
we have to teach customers. Upside is that we'll have to do that
eventually anyway, with the new collection-based traversal.