Allow application of enhancements without serialization (#550)
Prior to this PR, we called the EnhancementsRunner and DataFlowRunner in a few places where we needed enhancements/dataflow to be applied to the CPG, but we did not want them to be serialized into proto. We achieved this by simply passing a SerializedCpg without a filename. While this works, it means that the enhancementsRunner still serializes, it just does not write the serialized files anywhere. As serialization comes at a cost, with this PR, we now allow passing in no SerializedCpg at all, in which case the EnhancementRunner/DataflowRunner only applies but does not serialize the graph.

* Allow application of enhancements without serialization.

* Methods for backward compatibility.

* Same change for DataFlowRunner.

* Make `loadFromOverflowDb` less confusing.

* Update dataflowengine/src/main/scala/io/shiftleft/dataflowengine/layers/dataflows/DataFlowRunner.scala

Co-Authored-By: Michael Pollmeier <michael@michaelpollmeier.com>

* Revert "Update dataflowengine/src/main/scala/io/shiftleft/dataflowengine/layers/dataflows/DataFlowRunner.scala"

This reverts commit f84d2f3f79f6e6460a12e9743c0918ac4c1c9b87.

Co-authored-by: Michael Pollmeier <michael@michaelpollmeier.com>