comment.file: handle in NodeSteps for safety and clarity
Before this PR, whether we use the correct traversal depended on the
usage when calling `.file`. If we statically only know that we have a
`NodeSteps[nodes.StoredNodes]`, they will *not* use the more specific
`Comment.file`, even if the nodes all actually *are* of type
`nodes.Comment`.

Therefor it's safer (and easier to reason about) if we make the
distinction on NodeSteps. The obvious downside is that we clutter
`Comment` related things in `NodeSteps`, but that was already the case
for Namespace.