RFC: use typeclasses for pretty printing
We override `.p` in a few different steps, always relying on
overriding it in a more specific `Steps` class. This is rather clumsy
and doesn't fit well with our new streamlined implicit extension
design. This PR proposes to replace it with type classes. The tests
show all use cases we have for `.p`.

This is mostly a RFC PR. If we like this pattern, I'll
replace the old `.p` everywhere.

Related: https://github.com/ShiftLeftSecurity/codescience/issues/3406