rewrite and fix BindingMethodOverridesPass
Consider
```
class Base{
def foo():Int = 1
}
class Derived extends Base {}
```
In this case, the method `foo` is not overridden: The derived class `Derived` has a binding/vtable entry, which points to the same function as the entry in `Base`.

The old implementation of this pass incorrectly believed that this function is overridden. Due to a lack of unit tests, this was not caught. This causes extensive unneeded type-resolution attempts during dataflow tracking, with impact both on speed and precision.

Co-authored-by: Olof-Joachim Frahm (欧雅福) <olof@shiftleft.io>