return nodesteps instead of iterators (#771)
In an attempt to avoid unnecessary overhead I previously changed the
return type of `method.parameter` from `NodeSteps` to `Iterator`.
Given that it complicates things - the user now needs to call
`.start` explicitly, or we add a hard to understand implicit
conversion. Once we use overflowdb traversal, it won't have any
overhead any more, so, I'll change that back then.