Modify how contained_by works (#863)
This changes the internal representation of contained nodes. These behave, API-wise, like normal fields of nodes.

The old representation used CONTAINED_NODE edges with properties refering to "LOCAL_NAME", and sorted by "INDEX". This was slow in two ways: First, each access required a collection of outgoing edges, filtering by LOCAL_NAME and sorting by INDEX. Second, when storing a node that refers to another one, the old node needs to be modified (because its array of incoming edges grows).

The new representation uses normal fields. These are proto-serialized in a new property; and they use the pre-existing infrastructure for node-references in the overflowdb serialization. 

In order to accomodate this, diffgraph application has been reworked: We don't make a detour via string-keys anymore. Instead the NewNode and StoredNode subclasses are tightly coupled, and know each other. 

Further, this changes the API between overflowdb and codepropertygraph for setting higher cardinality properties. These are now viewed as list-valued properties. The old code took quadratic time in the length of these lists for deserialization of nodes. That was okay because we didn't have any very long lists in practice.  Now we model CONTAINED_NODE as list-valued properties, and these lists can be long.