Update scalatest to 3.1.1.
I need to change all the import because of an API change in scalatest.
This resolves a problem in Intellij is not able to run codescience unit
tests because of conflicting scalatest versions. We end up with
conflicting scalatest versions because currently semanticcpg used
scalatest as full dependency and not just as a test dependency.