Add TYPE_REF node.
TYPE_REF node is a generalisation of METHOD_REF.
For both the relevant property is TYPE_FULL_NAME which tells the data
flow tracker what methods can be dispatched on an expression.
In the METHOD_REF case this is usually only one METHOD which is
additionally indicated by its METHOD_FULL_NAME property.
TYPE_REF on the other hand stands for multiple METHODs which can be
dispatched an thus the METHOD_FULL_NAME property is no longer needed.

TYPE_REF is for now used to represent Javascript classes.
For constructs like: var x = class Foo(){}
we create a TYPE_REF for the right hand side of the assignment.