Cleanup (very) verbose logging in tests (#962)
Noticed that regular `sbt> test` produced a ton of logs that were hard
to follow and not even possible to configure. In order to improve the
situation:
1. Moved test fixtures code from `src/main` to `test`, since it is
really test code
2. This enabled me to make `fuzzyc2cpg` a test dependency rather than a
compile time dependency and forcing it on everyone
3. Excluded `logback` from all `fuzzyc2cpg` deps since for some unknown
reason it is the logging implementation of choice there. All other
projects only make log4j2 as a runtime dep, so someone should revisit
that
4. Added log4j2 config that is **only** used in tests (therefore not
being passed downstream)
5. Needed to fork tests in order to be able to assign the external
config, which brought its own problems when dealing with subprojects

As a result `sbt> test` produces much, much less noise and one is able
to control it via the `resources/log4j2-test.xml` file, if necessary.