refactor tracking point methods (#970)
This unifies the computation of `trackedBase` and `accessPath` from a tracking point. The underlying observation is that these are only needed as a pair, conceptually. The computation is now also available via an implicit, if it is explicitly imported.

Furthermore, the handling of codescience-specific special cases is partially moved here, via the FakeTrackingPoint interface. This is in order to make the related functions in codescience obsolete.

This change is in preparation for handling of aliases.