fix intellij / bsp build: remove the baseDirectory setting
This innocent-looking setting made intellij choke on the subprojects and didn't recognise them as scala projects properly.
Unfortunately this setting was necessary because intellij and sbt have different default working directories for executing tests.
This PR uses a hack to achieve the same: it looks for a file that (hopefully always) only be in the project root.
If the current working dir doesn't contain that file, we must be in a subproject, and will therefor navigate up one directory.

Turns out: this same technique was already applied in ScriptManagerTest.scala before - someone else has a knack for hacks

I ran all tests from within Intellij and it all seems to be green