don't handle schema extensions in plugin manager any more (#1187)
* don't handle schema extensions in plugin manager any more

This one may be controversial, so let me explain:
We used to define the schema in json, which meant we didn't have any
IDE and compiler support. Now that it's a Scala DSL we're in a much
better situation overall, however we cannot simply copy and merge
schema parts together by copying a json file. The schema needs to be
stitched together at compile time.

Now we could allow plugins to define their schema extension parts in
scala code snippets and copy them together during `./joern
--add-plugin` - this is pretty much what I wanted to achieve in
https://github.com/joernio/docker-ext/pull/9/ - the downside is that
we lose IDE/compiler support during plugin development...

tl;dr; there's different options/pros/cons, and we may find a better
model in the future, but for now I think this is the best compromise
with the new improved schema DSL.

* drop old test

* fix tests properly

* fmt