detect installation root dir via marker file (#1404)
* use current working dir for workspace

some users may want to install joern/ocular system-wide in a read-only
directory, and have their workspace e.g. in their user home
re https://discord.com/channels/832209896089976854/832214243230744626/887106423341875200

* make console/basedir configurable, mostly for tests

* detect installation root dir via marker file

https://github.com/ShiftLeftSecurity/codepropertygraph/pull/1387
introduced the assumption that the root directory is always two levels
up from the lib dir. That's not true for ocular, and a little brittle anyway.
This PR changes the behaviour to instead search for a marker file.

* adjust test