add parallelism disabling hook to base classes for CpgPass (#1748)
In order to parallelize a pass, it is useful during development to be able to use the ForkJoinParallelCpgPass framework, but without actually parallelizing. This can be done by overriding the new `isParallel` hook. This also allows to make parallelization configurable (e.g. if one is afraid of race-conditions and wants a feature-flag).